Refactor RegistrationForm to TS
